Move `Node` to its own module
