+++ src/data/syncset/mod.rs
@@ -4,10 +4,11 @@ mod syncset;
 mod syncerror;
 mod path;
 mod set;
+mod node;
 
 // Imports
 pub use set::Set;
 pub use syncset::SyncSet;
 pub use path::*;
 pub use syncerror::*;
-use syncset::Node;
+use node::Node;
